refactor(scheduler): scope least-loaded reservations
